screeing/requirements.txt
qdrant-client==1.8.0
fastapi==0.115.12
python-dotenv==1.0.0
google-generativeai==0.8.5
sentence-transformers==4.1.0
faiss-cpu==1.11.0
pandas==2.0.3
numpy==1.26.0
protobuf==3.19.6
